Java: Integrate EmbeddingVector into Embedding (#2328)
### Motivation and Context

<!-- Thank you for your contribution to the semantic-kernel repo!
Please help reviewers and future users, providing the following
information:
  1. Why is this change required?
  2. What problem does it solve?
  3. What scenario does it contribute to?
  4. If it fixes an open issue, please link to the issue here.
-->
EmbeddingVector and the vectoroperations package are down in
semantickernel-core, which makes it impossible to use the cosine
similarity function needed to calculate the similarity between
embeddings.
### Description

<!-- Describe your changes, the overall approach, the underlying design.
These notes will help understanding how your code works. Thanks! -->
EmbeddingVector was integrated into Embedding and the vectoroperations
package was refactored into a VectorOperations class which is in the
same package as Embedding. Now, one can do:
`embedding.cosineSimilarity(otherEmbedding)` instead of having to new up
an EmbeddingVector from embedding.getVector().

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [X] The code builds clean without any errors or warnings
- [X] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [X] All unit tests pass, and I have added new tests where possible
- [X] I didn't break anyone :smile:

+++ java/semantickernel-core/src/test/java/com/microsoft/semantickernel/memory/VolatileMemoryStoreTests.java
@@ -814,7 +814,7 @@ class VolatileMemoryStoreTests {
         for (Iterator<Tuple2<MemoryRecord, Float>> iterator = topNResults.iterator();
                 iterator.hasNext(); ) {
             Tuple2<MemoryRecord, Float> tuple = iterator.next();
-            int compare = Double.compare(tuple.getT2(), 0.75);
+            int compare = Float.compare(tuple.getT2(), 0.75f);
             assertTrue(topNKeys.contains(tuple.getT1().getKey()));
             assertTrue(compare >= 0);
         }
